Refactoring of schools.js
Extract marker content methods that are repeated.  Extract cluster group method that is repeated.  Move programsToggle on click to schools.js because using these common methods.  Add removeLayer calls to onclick of each of the toggles so it can reset the map.  Add variables for items on menu to top of js file to use in other parts -- includes clusterGroup variable which is changed at different times.  Some more refactoring needs to be done for the filtering of school types - this refactoring may be used for the programs.