Modify toggle for search.  Refactor filtering so all of it is under one toggle button.
Change createFilterList method to work for both types and programs.  Modify DisplayFilterList method to not add an event listener to the checkbox.  Add listener to divs with lists to call new method changeMap which updates the map according to filtered selection.  Still working on getting it so both filters work at once.  Need to remove code from previous filtering toggles.